Renamed lnst.Slave module to lnst.Agent
Moved all things under lnst.Slave to lnst.Agent

The files and some import statements where fixed in this commit, there might be more imports and other references that will be fixed in future commits.
